Fix signature error on RexxQueryExit
git-svn-id: https://svn.code.sf.net/p/oorexx/code-0/main/trunk@3563 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ rexxapi/client/RegistrationAPI.cpp
@@ -371,7 +371,7 @@ RexxReturnCode RexxEntry RexxDeregisterExit(
 RexxReturnCode RexxEntry RexxQueryExit(
   const char *      name,              /* Environment Name           */
   const char *      module,            /* Associated Name (of DLL)   */
-  size_t             *exist,           /* existence information      */
+  unsigned short     *exist,           /* existence information      */
   char               *userWord)        /* data from registration     */
 {
     *exist = 0;
